Move coresForNode() into cpu_linux.go
